+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -6,14 +6,6 @@
 //
 import Foundation
 
-public protocol FeedStore {
-    typealias DeletionCompletion = (Error?) -> Void
-    typealias InsertionCompletion = (Error?) -> Void
-
-    func deleteCachedFeed(completion: @escaping DeletionCompletion)
-    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
-}
-
 public final class LocalFeedLoader {
     
     private let store: FeedStore
